backend/requirements.txt
aiohttp==3.11.13
anthropic==0.71.0
celery==5.3.6
fastapi==0.120.0
motor==3.7.1
neo4j==6.0.2
numpy==2.3.4
openai==0.28.0
passlib==1.7.4
protobuf==6.33.0
pydantic==1.10.22
pymongo==4.15.3
python-dotenv==1.1.1
python-jose==3.3.0
pyttsx3==2.99
pywhatkit==5.4
redis==5.0.1
requests==2.32.3
sentence-transformers==2.5.0
SpeechRecognition==3.10.0
youtube-search-python==1.6.6
yt-dlp==2025.10.22
uvicorn==0.30.1
websockets==12.0
